further reorganize
remove unused isProxiedResult function

move unwrapResult and dehoistResult into createMergedResolver

WrapFields and HoistField transforms now use their own unwrapping and dehoisting logic, so these functions should be located only to the file that used them
